Add focused validator unit tests and remove outdated test blocks
Add new test file internal/runner/security/validator_test.go containing focused unit tests for DefaultConfig, NewValidator and NewValidatorWithFS.

Remove a large set of previously existing/duplicated validator test functions to simplify and consolidate test coverage.

Ensure tests use the project's test build tag and common testing helpers.
